+++ Packages/Core/Sources/Features/AuthFeature/LoginView.swift
@@ -1,6 +1,7 @@
 import ComposableArchitecture
 import Entity
 import SwiftUI
+import UIComponents
 
 public struct LoginView: View {
     @Bindable var store: StoreOf<LoginReducer>
@@ -90,61 +91,6 @@ extension LoginView {
             Text("メールアドレスとパスワードでログイン")
                 .multilineTextAlignment(.center)
         }
-    }
-}
-
-struct PrimaryButtonStyle: ButtonStyle {
-    @Environment(\.isEnabled) var isEnabled: Bool
-
-    func makeBody(configuration: Self.Configuration) -> some View {
-        HStack {
-            Spacer()
-            configuration.label
-            Spacer()
-        }
-            .bold()
-            .foregroundColor(isEnabled ? .white : Color(.placeholderText))
-            .padding(13)
-            .background(isEnabled ? .orange : Color(.secondarySystemFill))
-            .opacity(configuration.isPressed ? 0.2 : 1.0) // タップしている間は色を薄く
-            .clipShape(RoundedRectangle(cornerRadius: 16))
-            .hoverEffect()
-    }
-}
-
-extension ButtonStyle where Self == PrimaryButtonStyle {
-    static var primary: PrimaryButtonStyle {
-        .init()
-    }
-}
-
-struct RoundedBorderModifier<Style: ShapeStyle>: ViewModifier {
-    var style: Style, width: CGFloat = 0, radius: CGFloat
-
-    func body(content: Content) -> some View {
-        content
-            .overlay {
-                RoundedRectangle(cornerRadius: radius)
-                    .stroke(lineWidth: width*2)
-                    .fill(style)
-            }
-            .mask {
-                RoundedRectangle(cornerRadius: radius)
-            }
-    }
-}
-extension View {
-    func roundedBorder<S: ShapeStyle>(
-        _ style: S,
-        width: CGFloat,
-        radius: CGFloat
-    ) -> some View {
-        let modifier = RoundedBorderModifier(
-            style: style,
-            width: width,
-            radius: radius
-        )
-        return self.modifier(modifier)
     }
 }
